+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - Added `METADATA_REMOVE` event to facilitate removal of image layer metadata from the `DescriptionSubscriber` upon layer removal.
 - Added support for providing cell sets files in view configs.
 - Added support for responsive height in the `Welcome` and `PubSubVitessceGrid` components.
-
+- Added a prop `additionalCellScatterplotLayerProps` to the `Scatterplot` component.
 
 ### Changed
 - Refactored the Scatterplot and Spatial components. Removed the AbstractSelectableComponent class. Moved getter functions to props.